requirements.txt
yfinance==0.2.66
pandas>=2.3.3
pandas-datareader>=0.10.0
numpy>=2.2.6
feature-engine>=1.9.3
requests-cache>=1.2.1
scikit-learn>=1.6.1
wavetrainer>=0.3.6
tqdm>=4.67.1
pyvinecopulib>=0.7.5
fredapi>=0.5.2
python-dotenv>=1.1.0
kaleido>=1.2.0
plotly>=6.3.1
scipy>=1.16.3
joblib>=1.5.2
pyarrow>=22.0.0
alpaca-py>=0.43.2